+++ app/src/main/java/com/example/olpgas/browse_rooms/data/remote/model/AllRoomDetails.kt
@@ -14,7 +14,7 @@ data class AllRoomDetails(
     val deposit: Int,
     val city: String,
     val state: String,
-    val ratings: Int,
+    val ratings: Float,
     val bookingStatus: String,
     val features: List<String>
 )
