+++ app/src/main/java/com/example/olpgas/view_room_details/data/local/database/entities/FullRoomDetailsLocal.kt
@@ -24,6 +24,6 @@ data class FullRoomDetailsLocal(
     val description: String,
     val roomFeatureId: Int,
     val bookingStatus: String,
-    val ratings: Int,
+    val ratings: Float,
     val urls: List<String>
 )
