+++ app/src/main/java/com/example/olpgas/view_room_details/data/remote/model/FullRoomDetails.kt
@@ -22,5 +22,5 @@ data class FullRoomDetails(
     val description: String,
     val roomFeatureId: Int,
     val bookingStatus: String,
-    val ratings: Int
+    val ratings: Float
 )
